chore: move test doubles to separate folder
