+++ Modules/Quotas/test/Quotas.Application.Tests/TestDoubles/FindTiersStubRepository.cs
@@ -1,15 +1,15 @@
 ﻿using Backbone.Modules.Quotas.Application.Infrastructure.Persistence.Repository;
 using Backbone.Modules.Quotas.Domain.Aggregates.Tiers;
 
-namespace Backbone.Modules.Quotas.Application.Tests.Tests.Identities;
+namespace Backbone.Modules.Quotas.Application.Tests.TestDoubles;
 
-public class FindTierQuotaDefinitionsStubRepository : ITiersRepository
+public class FindTiersStubRepository : ITiersRepository
 {
-    private readonly TierQuotaDefinition _tierQuotaDefinition;
+    private readonly Tier _tier;
 
-    public FindTierQuotaDefinitionsStubRepository(TierQuotaDefinition tierQuotaDefinition)
+    public FindTiersStubRepository(Tier tier)
     {
-        _tierQuotaDefinition = tierQuotaDefinition;
+        _tier = tier;
     }
 
     public Task Add(Tier tier, CancellationToken cancellationToken)
@@ -19,12 +19,12 @@ public class FindTierQuotaDefinitionsStubRepository : ITiersRepository
 
     public Task<Tier> Find(string id, CancellationToken cancellationToken, bool track = false)
     {
-        throw new NotImplementedException();
+        return Task.FromResult(_tier);
     }
 
     public Task<TierQuotaDefinition> FindTierQuotaDefinition(string id, CancellationToken cancellationToken, bool track = false)
     {
-        return Task.FromResult(_tierQuotaDefinition);
+        throw new NotImplementedException();
     }
 
     public Task Update(Tier tier, CancellationToken cancellationToken)
