+++ d/register-for-push-notifications.puml
@@ -1,0 +1,38 @@
+@startuml Register for push notifications
+!pragma teoz true
+autonumber
+hide footbox
+autoactivate on
+
+participant "UpdateDeviceRegistrationCommand" as command
+participant "IPushService\n(DirectPushService)" as pushService
+participant "IPnsRegistrationRepository\n(PnsRegistrationRepository)" as repo
+participant "DevicesDbContext" as dbContext
+participant "PnsRegistration" as registration
+
+activate command
+command -> command : PnsHandle.Parse(command.Handle, command.Platform)
+return handle
+command -> pushService : UpdateRegistration(userContext.GetAddress(), userContext.getDeviceId(), handle)
+    pushService -> repo : GetByDeviceId(deviceId)
+        repo -> dbContext : PnsRegistrations.FirstOrDefault(deviceId)
+        return null
+    return null
+
+    create registration 
+    pushService -> registration : new(address, deviceId, handle)
+        autoactivate off
+        registration -> registration : IdentityAddress = address
+        registration -> registration : DeviceId = deviceId
+        registration -> registration : Handle = handle
+        registration -> registration : CreatedAt = SystemTime.UtcNow
+        autoactivate on
+    return registration
+    
+    pushService -> repo : Add(registration)
+        repo -> dbContext : PnsRegistrations.Add(registration)
+        return
+    return
+return
+deactivate command
+@enduml
