+++ d/update-registration-for-push-notifications.puml
@@ -1,0 +1,35 @@
+@startuml Update registration for push notifications
+!pragma teoz true
+autonumber
+hide footbox
+autoactivate on
+
+participant "UpdateDeviceRegistrationCommand" as command
+participant "IPushService\n(DirectPushService)" as pushService
+participant "IPnsRegistrationRepository\n(PnsRegistrationRepository)" as repo
+participant "DevicesDbContext" as dbContext
+participant "PnsRegistration" as registration
+
+activate command
+command -> command : PnsHandle.Parse(command.Handle, command.Platform)
+return handle
+command -> pushService : UpdateRegistration(userContext.GetAddress(), userContext.getDeviceId(), handle)
+    pushService -> repo : GetByDeviceId(deviceId)
+        repo -> dbContext : PnsRegistrations.FirstOrDefault(deviceId)
+        return registration
+    return registration
+
+    pushService -> registration : Update(handle)
+        autoactivate off
+        registration -> registration : Handle = handle
+        registration -> registration : UpdatedAt = SystemTime.UtcNow
+        autoactivate on
+    return
+    
+    pushService -> repo : Update(registration)
+        repo -> dbContext : PnsRegistrations.Update(registration)
+        return
+    return
+return
+deactivate command
+@enduml
